+++ backend/src/controllers/project/addMemberToProject.ts
@@ -1,0 +1,22 @@
+import { Request, Response } from "express";
+
+import projectService from "../../services/projectService";
+import { isProjectDeleted } from "../../utils/projectUtils";
+import { PROJECT_NOT_FOUND_ERROR } from "./errors";
+
+export default async function addMemberToProject(
+  request: Request,
+  response: Response,
+): Promise<void> {
+  const { projectId } = request.params;
+  const { userId } = request.body;
+
+  const project = await projectService.getProject(projectId);
+
+  if (isProjectDeleted(project)) {
+    throw PROJECT_NOT_FOUND_ERROR;
+  }
+
+  await projectService.addMembersToProject([userId], projectId);
+  response.sendStatus(204);
+}
