+++ backend/src/controllers/user/getUserById.ts
@@ -2,6 +2,7 @@ import { Request, Response } from "express";
 import { param, ValidationChain } from "express-validator";
 
 import UserService from "../../services/UserService";
+import { wrapResponse } from "../../utils/responseUtils";
 
 export const getUserByIdValidationChains: ValidationChain[] = [
   param("userId").isMongoId().notEmpty(),
@@ -13,5 +14,5 @@ export default async function getUserById(request: Request, response: Response):
 
   const user = await userService.fetchUserById(userId);
 
-  response.send({ data: user });
+  response.send(wrapResponse(user));
 }
