+++ backend/src/controllers/user/getUsers.ts
@@ -1,6 +1,7 @@
 import { Request, Response } from "express";
 
 import UserService from "../../services/UserService";
+import { wrapResponse } from "../../utils/responseUtils";
 
 export default async function getUsers(request: Request, response: Response): Promise<void> {
   const { q } = request.query;
@@ -8,5 +9,5 @@ export default async function getUsers(request: Request, response: Response): Pr
 
   const users = await userService.fetchUsers(q as string);
 
-  response.send({ data: users });
+  response.send(wrapResponse(users));
 }
